MAGECLOUD-1020: [develop] Enterprise Cloud Edition Sets Up an Admin User with A Default Password
