+++ src/Test/Unit/Process/Deploy/InstallUpdate/Update/AdminCredentialsTest.php
@@ -120,12 +120,14 @@ class AdminCredentialsTest extends TestCase
 
         $query = 'UPDATE `admin_user` SET `lastname` = ? ORDER BY `user_id` ASC LIMIT 1';
 
-        $this->loggerMock->expects($this->exactly(3))
+        $this->loggerMock->expects($this->once())
             ->method('info')
+            ->with('Updating admin credentials.');
+        $this->loggerMock->expects($this->exactly(2))
+            ->method('warning')
             ->withConsecutive(
                 ['Some administrator already uses this email ' . $email],
-                ['Some administrator already uses this username ' . $userName],
-                ['Updating admin credentials.']
+                ['Some administrator already uses this username ' . $userName]
             );
         $this->environmentMock->expects($this->once())
             ->method('getAdminPassword')
